apps/qlib-training-data/requirements.txt
# Qlib Training Data Generation App - Requirements

# Core dependencies
pyqlib>=0.9.0
pandas>=1.3.0
numpy>=1.20.0

# Optional but recommended
scikit-learn>=1.0.0
matplotlib>=3.3.0
seaborn>=0.11.0

# For data storage
pyarrow>=6.0.0  # For parquet support
